Make ModelInfoLookup into an interface
