+++ src/main/java/com/nedap/archie/archetypevalidator/ArchetypeValidator.java
@@ -9,7 +9,7 @@ import com.nedap.archie.archetypevalidator.validations.NodeIdValidation;
 import com.nedap.archie.archetypevalidator.validations.TerminologyValidation;
 import com.nedap.archie.archetypevalidator.validations.ValueSetValidation;
 import com.nedap.archie.rminfo.ArchieRMInfoLookup;
-import com.nedap.archie.rminfo.ModelInfoLookup;
+import com.nedap.archie.rminfo.ReflectionModelInfoLookup;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -23,13 +23,13 @@ public class ArchetypeValidator {
     private static final Logger logger = LoggerFactory.getLogger(ArchetypeValidator.class);
 
     private List<ArchetypeValidation> validations;
-    private ModelInfoLookup modelInfoLookup;
+    private ReflectionModelInfoLookup modelInfoLookup;
 
     public ArchetypeValidator() {
         this(ArchieRMInfoLookup.getInstance());
     }
 
-    public ArchetypeValidator(ModelInfoLookup lookup) {
+    public ArchetypeValidator(ReflectionModelInfoLookup lookup) {
         modelInfoLookup = lookup;
 
         validations = new ArrayList<>();
