+++ src/main/java/com/nedap/archie/archetypevalidator/validations/ModelConformanceValidation.java
@@ -7,7 +7,7 @@ import com.nedap.archie.aom.CPrimitiveObject;
 import com.nedap.archie.archetypevalidator.ErrorType;
 import com.nedap.archie.archetypevalidator.ValidatingVisitor;
 import com.nedap.archie.archetypevalidator.ValidationMessage;
-import com.nedap.archie.rminfo.ModelInfoLookup;
+import com.nedap.archie.rminfo.ReflectionModelInfoLookup;
 import com.nedap.archie.rminfo.RMAttributeInfo;
 import com.nedap.archie.rminfo.RMTypeInfo;
 
@@ -20,7 +20,7 @@ import java.util.List;
  */
 public class ModelConformanceValidation extends ValidatingVisitor {
 
-    public ModelConformanceValidation(ModelInfoLookup lookup) {
+    public ModelConformanceValidation(ReflectionModelInfoLookup lookup) {
         super(lookup);
     }
