+++ src/test/java/com/nedap/archie/adlparser/APathQueryRMTest.java
@@ -7,7 +7,7 @@ import com.nedap.archie.rm.archetyped.Pathable;
 import com.nedap.archie.rm.composition.Composition;
 import com.nedap.archie.rm.datavalues.DvText;
 import com.nedap.archie.rminfo.ArchieRMInfoLookup;
-import com.nedap.archie.rminfo.ModelInfoLookup;
+import com.nedap.archie.rminfo.ReflectionModelInfoLookup;
 import com.nedap.archie.testutil.TestUtil;
 import org.junit.Before;
 import org.junit.Test;
@@ -69,7 +69,7 @@ public class APathQueryRMTest {
             composition.getContext().getOtherContext().getItems().addAll(composition2.getContext().getOtherContext().getItems());
         }
 
-        ModelInfoLookup modelInfoLookup = ArchieRMInfoLookup.getInstance();
+        ReflectionModelInfoLookup modelInfoLookup = ArchieRMInfoLookup.getInstance();
 
         List<RMObjectWithPath> context = new APathQuery("/context")
                 .findList(modelInfoLookup, composition);
